# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Node Comics API",
    "description": "私有漫画图片、持久化翻译任务、普通与 PLUS 会员权益及周期页数额度。",
    "version": "0.3.0"
  },
  "paths": {
...
    "/v1/translation-submissions": {
      "post": {
        "tags": [
          "submissions"
        ],
        "summary": "Submit",
        "operationId": "submit_v1_translation_submissions_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "idempotency-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmissionRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "submissions"
        ],
        "summary": "Submissions",
        "operationId": "submissions_v1_translation_submissions_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 30,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/auth/dev": {
      "post": {
        "summary": "Dev Login",
        "operationId": "dev_login_v1_auth_dev_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DevLogin"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/capabilities": {
      "get": {
        "summary": "Capabilities",
        "operationId": "capabilities_v1_capabilities_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CapabilitiesResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/v1/me/usage": {
      "get": {
        "summary": "Usage",
        "operationId": "usage_v1_me_usage_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 30,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CapabilitiesResponse": {
        "properties": {
          "modes": {
            "items": {
              "$ref": "#/components/schemas/ModeCapability"
            },
            "type": "array",
            "title": "Modes"
          },
          "languages": {
            "items": {
              "$ref": "#/components/schemas/LanguageCapability"
            },
            "type": "array",
            "title": "Languages"
          },
          "limits": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Limits"
          },
          "entitlements": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EntitlementsResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "retention_days": {
            "type": "integer",
            "title": "Retention Days"
          },
          "unknown_release_seconds": {
            "type": "integer",
            "title": "Unknown Release Seconds"
          }
        },
        "type": "object",
        "required": [
          "modes",
          "languages",
          "limits",
          "entitlements",
          "retention_days",
          "unknown_release_seconds"
        ],
        "title": "CapabilitiesResponse"
      },
...
      "DevLogin": {
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 60,
            "minLength": 1,
            "pattern": "^[\\w\\-\\u4e00-\\u9fff ]+$",
            "title": "Username"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "username"
        ],
        "title": "DevLogin"
      },
      "EntitlementsResponse": {
        "properties": {
          "plan": {
            "type": "string",
            "title": "Plan"
          },
          "plus_started_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plus Started At"
          },
          "plus_expires_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plus Expires At"
          },
          "timezone": {
            "type": "string",
            "title": "Timezone"
          },
          "queue_capacity": {
            "type": "integer",
            "title": "Queue Capacity"
          },
          "realtime_slots": {
            "type": "integer",
            "title": "Realtime Slots"
          },
          "scheduler_weight": {
            "type": "number",
            "title": "Scheduler Weight"
          },
          "modes": {
            "additionalProperties": {
              "$ref": "#/components/schemas/ModeEntitlement"
            },
            "type": "object",
            "title": "Modes"
          },
          "generated_at": {
            "type": "string",
            "title": "Generated At"
          },
          "pending_previous_period_pages": {
            "type": "integer",
            "title": "Pending Previous Period Pages"
          }
        },
        "type": "object",
        "required": [
          "plan",
          "plus_started_at",
          "plus_expires_at",
          "timezone",
          "queue_capacity",
          "realtime_slots",
          "scheduler_weight",
          "modes",
          "generated_at",
          "pending_previous_period_pages"
        ],
        "title": "EntitlementsResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ItemRequest": {
        "properties": {
          "client_item_id": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Client Item Id"
          },
          "image_sha256": {
            "type": "string",
            "pattern": "^[a-f0-9]{64}$",
            "title": "Image Sha256"
          },
          "byte_size": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Byte Size"
          },
          "content_type": {
            "type": "string",
            "maxLength": 60,
            "title": "Content Type"
          },
          "name": {
            "type": "string",
            "maxLength": 255,
            "title": "Name",
            "default": ""
          },
          "file_hash": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[a-f0-9]{64}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Hash"
          },
          "page_index": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 1000000.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Index"
          },
          "asset_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Asset Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "client_item_id",
          "image_sha256",
          "byte_size",
          "content_type"
        ],
        "title": "ItemRequest"
      },
...
      "LanguageCapability": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "label": {
            "type": "string",
            "title": "Label"
          }
        },
        "type": "object",
        "required": [
          "id",
          "label"
        ],
        "title": "LanguageCapability"
      },
...
      "LoginResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type"
          },
          "expires_in": {
            "type": "integer",
            "title": "Expires In"
          },
          "user": {
            "$ref": "#/components/schemas/UserResponse"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "token_type",
          "expires_in",
          "user"
        ],
        "title": "LoginResponse"
      },
...
      "ModeCapability": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          },
          "languages": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Languages"
          }
        },
        "type": "object",
        "required": [
          "id",
          "label",
          "enabled",
          "languages"
        ],
        "title": "ModeCapability"
      },
      "ModeEntitlement": {
        "properties": {
          "allowed": {
            "type": "boolean",
            "title": "Allowed"
          },
          "unlimited": {
            "type": "boolean",
            "title": "Unlimited"
          },
          "quota_kind": {
            "type": "string",
            "title": "Quota Kind"
          },
          "consent_version": {
            "type": "string",
            "title": "Consent Version"
          },
          "quota": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PeriodResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "allowed",
          "unlimited",
          "quota_kind",
          "consent_version",
          "quota"
        ],
        "title": "ModeEntitlement"
      },
...
      "PeriodResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "granted": {
            "type": "integer",
            "title": "Granted"
          },
          "used": {
            "type": "integer",
            "title": "Used"
          },
          "reserved": {
            "type": "integer",
            "title": "Reserved"
          },
          "available": {
            "type": "integer",
            "title": "Available"
          },
          "starts_at": {
            "type": "string",
            "title": "Starts At"
          },
          "resets_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resets At"
          },
          "next_expiry_at": {
            "type": "string",
            "title": "Next Expiry At"
          },
          "buckets": {
            "items": {
              "$ref": "#/components/schemas/QuotaBucketResponse"
            },
            "type": "array",
            "title": "Buckets"
          }
        },
        "type": "object",
        "required": [
          "id",
          "kind",
          "granted",
          "used",
          "reserved",
          "available",
          "starts_at",
          "resets_at",
          "next_expiry_at",
          "buckets"
        ],
        "title": "PeriodResponse"
      },
...
      "QuotaBucketResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "mode": {
            "type": "string",
            "title": "Mode"
          },
          "granted": {
            "type": "integer",
            "title": "Granted"
          },
          "used": {
            "type": "integer",
            "title": "Used"
          },
          "reserved": {
            "type": "integer",
            "title": "Reserved"
          },
          "available": {
            "type": "integer",
            "title": "Available"
          },
          "starts_at": {
            "type": "string",
            "title": "Starts At"
          },
          "expires_at": {
            "type": "string",
            "title": "Expires At"
          },
          "grants_access": {
            "type": "boolean",
            "title": "Grants Access"
          },
          "note": {
            "type": "string",
            "title": "Note"
          }
        },
        "type": "object",
        "required": [
          "id",
          "kind",
          "source",
          "mode",
          "granted",
          "used",
          "reserved",
          "available",
          "starts_at",
          "expires_at",
          "grants_access",
          "note"
        ],
        "title": "QuotaBucketResponse"
      },
...
      "SubmissionRequest": {
        "properties": {
          "mode": {
            "type": "string",
            "enum": [
              "classic",
              "redraw"
            ],
            "title": "Mode"
          },
          "target_language": {
            "type": "string",
            "title": "Target Language"
          },
          "max_quota_pages": {
            "type": "integer",
            "maximum": 1000000.0,
            "minimum": 0.0,
            "title": "Max Quota Pages"
          },
          "expected_kind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Kind"
          },
          "regenerate": {
            "type": "boolean",
            "title": "Regenerate",
            "default": false
          },
          "rerun_job_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rerun Job Id"
          },
          "acknowledge_unknown_cost": {
            "type": "boolean",
            "title": "Acknowledge Unknown Cost",
            "default": false
          },
          "reading_session_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 80
              },
              {
                "type": "null"
              }
            ],
            "title": "Reading Session Id"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ItemRequest"
            },
            "type": "array",
            "maxItems": 500,
            "minItems": 1,
            "title": "Items"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "mode",
          "target_language",
          "max_quota_pages",
          "items"
        ],
        "title": "SubmissionRequest"
      },
...
      "UsageEntry": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "job_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "period_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Period Id"
          },
          "quota_kind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quota Kind"
          },
          "pages": {
            "type": "integer",
            "title": "Pages"
          },
          "note": {
            "type": "string",
            "title": "Note"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "job_id",
          "kind",
          "period_id",
          "quota_kind",
          "pages",
          "note",
          "created_at"
        ],
        "title": "UsageEntry"
      },
      "UsageResponse": {
        "properties": {
          "entitlements": {
            "$ref": "#/components/schemas/EntitlementsResponse"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/UsageEntry"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "next_offset": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Offset"
          }
        },
        "type": "object",
        "required": [
          "entitlements",
          "items",
          "total",
          "next_offset"
        ],
        "title": "UsageResponse"
      },
...
      "UserResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "role": {
            "type": "string",
            "title": "Role"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "role"
        ],
        "title": "UserResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}